[SQL-Gen] Move ColType to its own file
Scala files should have only one class/trait/object and optional
companion object
