+++ src/main/scala/temple/generate/database/ast/ColType.scala
@@ -1,0 +1,13 @@
+package temple.generate.database.ast
+
+sealed trait ColType
+
+object ColType {
+  case object IntCol        extends ColType
+  case object FloatCol      extends ColType
+  case object StringCol     extends ColType
+  case object BoolCol       extends ColType
+  case object DateCol       extends ColType
+  case object TimeCol       extends ColType
+  case object DateTimeTzCol extends ColType
+}
